Refactor: move methods to ability.cpp
