+++ variables.hpp
@@ -812,8 +812,6 @@ int roll(int, int, int);
 int roll_max(int, int, int);
 int roundmargin(int = 0, int = 0);
 int route_info(int&, int&, int = 0);
-int skillexp(int = 0, int = 0, int = 0, int = 0, int = 0);
-int skillmod(int = 0, int = 0, int = 0);
 int talk_conv(std::string&, int = 0);
 int winposy(int = 0, int = 0);
 int zentohan(const std::string&, std::string&, int = 0);
@@ -982,23 +980,7 @@ void label_1346();
 void end_dmghp();
 void initialize_server_info();
 void label_1399();
-void label_1457();
-void label_1458();
-void label_1459(int);
-void gain_skill_experience_lock_picking(int);
-void gain_skill_experience_detection(int);
-void gain_skill_experience_casting(int);
-void gain_skill_experience_mana_capacity(int);
-void label_1464(int);
-void label_1465(int);
-void label_1466(int);
-void label_1468(int);
-void label_1469(int);
-void label_1470(int);
-void label_1471(int);
-void label_1472(int);
 void label_1512(int);
-void gain_disarm_trap_experience();
 void get_hungry(int);
 void label_1520(int);
 void refresh_burden_state();
